backend/requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-simple-history==3.10.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
qrcode==8.2
redis==6.3.0
reportlab==4.4.3
sentry-sdk[django]==2.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
